+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## Unreleased]
+### Added
+- check-xmpp-login.rb (@ushis)
 
 ## [0.0.3] - 2015-07-14
 ### Changed